Make cursors nullable to cater for empty array
Currently the schema does not allow for null cursors on empty arrays, and fails when the array is empty. 

I am using this as a baseline in a different repository, and don't currently use prisma2. I am not 100% sure what prisma returns in this case and cannot test it easily.

Will see if I can find typings online, but putting this PR in so long since it might be useful in the Prisma case as well.